Simplify and use JSX for class impl of table-app
